Function twilight_lavalink::http::get_route_planner

pub fn get_route_planner(
    address: SocketAddr,
    authorization: impl AsRef<str>
) -> Result<Request<&'static [u8]>, HttpError>

Available on crate feature http-support only.

Get the configured route planner for a node by address.

The response will include a body which can be deserialized into a RoutePlanner.

Errors

See the documentation for http::Error.