Function twilight_lavalink::http::load_track

pub fn load_track(
    address: SocketAddr, 
    identifier: impl AsRef<str>, 
    authorization: impl AsRef<str>
) -> Result<Request<&'static [u8]>, HttpError>

Get a list of tracks that match an identifier.

The response will include a body which can be deserialized into a LoadedTracks.