Module twilight_lavalink::model

Models to (de)serialize incoming/outgoing websocket events and HTTP responses.

Structs

Destroy
Destroy a player from a node.

Equalizer
Equalize a player.

EqualizerBand
A band of the equalizer event.

Pause
Pause or unpause a player.

Play
Play a track, optionally specifying to not skip the current track.

PlayerUpdate
An update about the information of a player.

PlayerUpdateState
New statistics about a node and its host.

Seek
Seek a player's active track to a new position.

SlimVoiceServerUpdate
A slimmed version of a twilight voice server update.

Stats
Statistics about a node and its host.

StatsCpu
CPU information about a node and its host.

StatsFrames
CPU information about a node and its host.

StatsMemory
Memory information about a node and its host.

Stop
Stop a player.

TrackEnd
A track ended.

TrackStart
A track started.

VoiceUpdate
A combined voice server and voice state update.

Volume
Set the volume of a player.

Enums

IncomingEvent
An incoming event from a Lavalink node.

Opcode
The type of event that something is.

OutgoingEvent
An outgoing event to send to Lavalink.

TrackEventType
The type of track event that was received.