Module twilight_lavalink::model::outgoing

Events that clients send to Lavalink.

Structs

Destroy
Destroy a player from a node.

Equalizer
Equalize a player.

EqualizerBand
A band of the equalizer event.

Pause
Pause or unpause a player.

Play
Play a track, optionally specifying to not skip the current track.

Seek
Seek a player's active track to a new position.

SlimVoiceServerUpdate
A slimmed version of a twilight voice server update.

Stop
Stop a player.

VoiceUpdate
A combined voice server and voice state update.

Volume
Set the volume of a player.

Enums

OutgoingEvent
An outgoing event to send to Lavalink.