Struct twilight_lavalink::node::Node

pub struct Node { /* private fields */ }

A connection to a single Lavalink server. It receives events and forwards events from players to the server.

Please refer to the module documentation.

Implementations

impl Node

pub async fn connect(
config: NodeConfig,
players: PlayerManager
) -> Result<(Self, IncomingEvents), NodeError>

Connect to a node, providing a player manager so that the node can update player details.

Please refer to the module documentation for some additional information about directly creating and using nodes. You are encouraged to use the Lavalink client instead.

Errors

Returns an error of type Connecting if the connection fails after several backoff attempts.

Returns an error of type BuildingConnectionRequest if the request failed to build.

Returns an error of type Unauthorized if the supplied authorization is rejected by the node.

pub const fn config(&self) -> &NodeConfig

Retrieve an immutable reference to the node’s configuration.

pub const fn players(&self) -> &PlayerManager

Retrieve an immutable reference to the player manager used by the node.

pub fn send(&self, event: OutgoingEvent) -> Result<(), NodeSenderError>

Retrieve an immutable reference to the node’s configuration.

Note that sending player events through the node’s sender won’t update player states, such as whether it’s paused.

Errors

Returns a NodeSenderErrorType::Sending error type if node is no longer connected.

pub fn sender(&self) -> NodeSender

Retrieve a unique sender to send events to the Lavalink server.

Note that sending player events through the node’s sender won’t update player states, such as whether it’s paused.

pub async fn stats(&self) -> Stats

Retrieve a copy of the node’s stats.

pub async fn penalty(&self) -> i32

Retrieve the calculated penalty score of the node.

This score can be used to calculate how loaded the server is. A higher number means it is more heavily loaded.