Struct twilight_lavalink::player::Player

pub struct Player { /* fields omitted */ }

A player for a guild connected to a node.

This can be used to send events over a node and to read the details of a player for a guild.

Implementations

impl Player

pub fn send(
&self,
event: impl Into<OutgoingEvent>
) -> Result<(), TrySendError<OutgoingEvent>>

Send an event to the player’s node.

Returns a futures_channel TrySendError if the node has been removed.

Examples

Send a Play and Pause event:

use twilight_lavalink::{model::{Play, Pause}, Lavalink};

let lavalink = Lavalink::new(user_id, 10);
let players = lavalink.players();

if let Some(player) = players.get(&guild_id) {
    player.send(Play::from((guild_id, track)))?;
    player.send(Pause::from((guild_id, true)))?;
}

pub fn node(&self) -> &Node

Return an immutable reference to the node linked to the player.

pub fn channel_id(&self) -> Option<ChannelId>

Return a copy of the player’s channel ID.

pub fn guild_id(&self) -> GuildId

Return an copy of the player’s guild ID.

pub fn paused(&self) -> bool

Return a copy of whether the player is paused.

pub fn position(&self) -> i64

Return a copy of the player’s position.

pub fn time_ref(&self) -> i64

Return a copy of the player’s time.

pub fn volume_ref(&self) -> u16

Return a copy of the player’s volume.