Module twilight_lavalink::http

Models to deserialize responses into and functions to create http crate requests.

Structs

FailingAddress
A failing IP address within the planner.

IpBlock
A block of IP addresses.

LoadedTracks
Possible track results for a query.

NanoIpDetails
Information about a Nano IP planner.

NanoIpRoutePlanner
A Nano IP planner.

PlaylistInfo
Information about a playlist from a search result.

RotatingIpDetails
Information about a Rotating IP planner.

RotatingIpRoutePlanner
A Rotating IP planner.

RotatingNanoIpDetails
Information about a Rotating Nano IP planner.

RotatingNanoIpRoutePlanner
A Rotating Nano IP planner.

Track
A track within a search result.

TrackInfo
Additional information about a track, such as the author.

Enums

IpBlockType
The IP version in use by the block.

LoadType
The type of search result given.

RoutePlanner
The route planner in use.

RoutePlannerType
The type of route planner in use.

Functions

get_route_planner
Get the configured route planner for a node by address.

load_track
Get a list of tracks that match an identifier.

unmark_failed_address
Unmark an IP address as being failed, meaning that it can be used again.