Function twilight_lavalink::http::unmark_failed_address

pub fn unmark_failed_address(
    node_address: impl Into<SocketAddr>, 
    authorization: impl AsRef<str>, 
    route_address: impl Into<IpAddr>
) -> Result<Request<Vec<u8>>, HttpError>

Unmark an IP address as being failed, meaning that it can be used again.

The response will not include a body on success.